web_parsing/requirements.txt
psycopg2-binary
SQLAlchemy
pandas
schedule
validators
requests
beautifulsoup4
lxml
scikit-learn